+++ modules/crawleradmin/Module.php
@@ -9,14 +9,17 @@ class Module extends \admin\base\Module
     public $apis = [
         'api-crawler-builderindex' => 'crawleradmin\apis\BuilderindexController',
         'api-crawler-index' => 'crawleradmin\apis\IndexController',
+        'api-crawler-searchdata' => 'crawleradmin\apis\SearchdataController',
     ];
 
     public function getMenu()
     {
-        return $this->node(Module::t('crawler'), 'pageview')
+        return $this->node(Module::t('crawler'), 'find_in_page')
             ->group(Module::t('crawler_indexing'))
-                ->itemApi(Module::t('crawler_index'), 'crawleradmin-index-index', 'visibility', 'api-crawler-index')
-                ->itemApi(Module::t('crawler_builderindex'), 'crawleradmin-builderindex-index', 'visibility_off', 'api-crawler-builderindex')
+                ->itemApi(Module::t('crawler_index'), 'crawleradmin-index-index', 'list', 'api-crawler-index')
+                //->itemApi(Module::t('crawler_builderindex'), 'crawleradmin-builderindex-index', 'visibility_off', 'api-crawler-builderindex')
+            ->group('Anylatics')
+            ->itemApi('Searchdata', 'crawleradmin-searchdata-index', 'label', 'api-crawler-searchdata')
         ->menu();
     }